EMAIL_SECURITY_SYSTEM_2/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
joblib==1.3.2
imaplib2==3.6
cryptography==41.0.4
psutil==5.9.5
python-dotenv==1.0.0